Refactor project model to calculate progress dynamically and update related components

+++ lib/features/home/presentation/widgets/in_progress_todo.dart
@@ -63,7 +63,7 @@ class InProgressTodo extends StatelessWidget {
                 disabledActiveTrackColor: Color(data.category.colorValue),
               ),
               child: Slider(
-                value: data.progress.toDouble() / 100,
+                value: data.progress / 100,
                 onChanged: null,
               ),
             ),
