+++ lib/features/project/presentation/manager/add_project_cubit/add_project_state.dart
@@ -33,7 +33,7 @@ final class AddProjectFormState extends AddProjectState {
   DateTime get startDate => project.startDate;
   DateTime get endDate => project.endDate;
   DateTime get date => project.startDate;
-  int get progress => project.progress;
+  double get progress => project.progress;
   TaskStatus get status => project.status;
   List<TaskModel> get tasks => project.tasks;
 }
